# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenCatalogi",
    "version": "2.1.16-unstable.20260919082905",
    "description": "OpenCatalogi's public, anonymously-reachable API surface: full-text search, catalog and publication read access, DCAT-AP-NL and schema.org harvest feeds, DIWOO sitemaps/robots.txt, the federation read surface, the federated-directory sync endpoint, and the CMS read APIs (pages/menus/themes/glossary/listings). Every path in this document corresponds to a route in `appinfo/routes.php` whose controller method carries `@PublicPage` / `#[PublicPage]` — admin-gated and authenticated-only endpoints (settings, retention, Woo batches, setup, OOAPI 5.0 resource reads, DCAT/DIWOO/OOAPI validation reports) are out of scope; see the `public-api-openapi-document` OpenSpec change.",
    "license": {
      "name": "EUPL-1.2",
      "url": "https://eupl.eu/1.2/en/"
    },
    "contact": {
      "name": "Conduction",
      "url": "https://www.conduction.nl/",
      "email": "info@conduction.nl"
    }
  },
  "servers": [
    {
      "url": "https://{host}/index.php/apps/opencatalogi",
      "description": "OpenCatalogi app base path on a Nextcloud instance.",
      "variables": {
        "host": {
          "default": "your-nextcloud.example.org",
          "description": "Hostname of the Nextcloud instance running OpenCatalogi."
        }
      }
    }
  ],
  "security": [],
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Standard error envelope. Some endpoints use `error`, some `message`, a few both.",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "DcatDocument": {
        "description": "DCAT-AP-NL RDF document. Serialization is content-negotiated via `?format=` (authoritative) or the `Accept` header; body shape depends on the negotiated media type (JSON-LD object, or a Turtle/RDF-XML string)."
      }
    },
    "parameters": {
...
      "FormatParam": {
        "name": "format",
        "in": "query",
        "description": "Explicit DCAT serialization format; overrides `Accept` header negotiation.",
        "schema": {
          "type": "string",
          "enum": [
            "jsonld",
            "turtle",
            "rdfxml"
          ]
        }
      },
      "DcatPageParam": {
        "name": "page",
        "in": "query",
        "description": "1-based page number for harvester-grade pagination of the per-catalog DCAT feed.",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
...
      "CatalogSlugPath": {
        "name": "catalogSlug",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9-]+$"
        }
      },
...
    },
    "headers": {
      "AccessControlAllowOrigin": {
        "description": "CORS allowed origin (`*` unless an allowlist is configured).",
        "schema": {
          "type": "string"
        }
      },
      "AccessControlAllowMethods": {
        "description": "CORS allowed HTTP methods.",
        "schema": {
          "type": "string"
        }
      },
      "AccessControlAllowHeaders": {
        "description": "CORS allowed request headers.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "The requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "CorsPreflight": {
        "description": "Preflight answered; no body.",
        "headers": {
          "Access-Control-Allow-Origin": {
            "$ref": "#/components/headers/AccessControlAllowOrigin"
          },
          "Access-Control-Allow-Methods": {
            "$ref": "#/components/headers/AccessControlAllowMethods"
          },
          "Access-Control-Allow-Headers": {
            "$ref": "#/components/headers/AccessControlAllowHeaders"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/catalogs/{catalogSlug}/dcat": {
      "get": {
        "operationId": "dcat-catalog",
        "summary": "Per-catalog DCAT-AP-NL document",
        "description": "DCAT-AP-NL document for one catalog. Only publicly-visible objects appear (DCAT-003). Supports conditional GET via ETag (304) and harvester-grade pagination via `page` (DCAT-008).",
        "tags": [
          "DCAT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CatalogSlugPath"
          },
          {
            "$ref": "#/components/parameters/FormatParam"
          },
          {
            "$ref": "#/components/parameters/DcatPageParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/DcatDocument"
                }
              },
              "text/turtle": {
                "schema": {
                  "$ref": "#/components/schemas/DcatDocument"
                }
              },
              "application/rdf+xml": {
                "schema": {
                  "$ref": "#/components/schemas/DcatDocument"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (ETag matched `If-None-Match`)."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "description": "Unsupported format requested."
          }
        }
      },
      "options": {
        "operationId": "dcat-catalog-preflight",
        "summary": "CORS preflight",
        "description": "CORS preflight for the per-catalog DCAT document.",
        "tags": [
          "DCAT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CatalogSlugPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CorsPreflight"
          }
        }
      }
    },
...
  }
}